Reorg of lib folder and some minor refactors
