+++ src/controls.tsx
@@ -21,7 +21,9 @@ import VectorInput from '@playcanvas/pcui/VectorInput/component';
 import SelectInput from '@playcanvas/pcui/SelectInput/component';
 // @ts-ignore: library file import
 import BindingTwoWay from '@playcanvas/pcui/BindingTwoWay';
+
 import React, { useEffect, useState, useContext } from 'react';
+
 import { Morph, Option, HierarchyNode } from './types';
 
 const ObserverContext = React.createContext(null);
